remove container and optimize initialization time
