+++ pyproject.toml
@@ -5,7 +5,6 @@ build-backend = "hatchling.build"
 [tool.hatch.build.targets.sdist]
 exclude = [
   "/.github",
-  "/docker",
   "/docs",
   "/images",
   "/scripts",
